requirements.txt
asgiref==3.5.2
astroid==2.12.12
certifi==2022.9.24
charset-normalizer==2.1.1
cmake==3.25.0
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.6
Django==4.1.3
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
install==1.3.5
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.8.0
logilab-common==1.9.7
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
openapi-codec==1.3.2
platformdirs==2.5.4
PyJWT==2.6.0
pylint==2.15.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
PyMySQL==1.0.2
pytz==2022.6
requests==2.28.1
simplejson==3.18.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
